+++ backend/server.js
@@ -23,6 +23,7 @@ const raisonRoutes = require('./src/routes/raisonRoutes')
 const serviceRoutes = require('./src/routes/serviceRoutes') 
 const loginRoutes = require('./src/routes/loginRoutes')
 const statsRoutes = require('./src/routes/stats/statsRoutes')
+const exportRoutes = require('./src/routes/exports/exportRoutes')
 
 app
 .use(serviceRoutes)
@@ -30,6 +31,7 @@ app
 .use(visiteRoutes)
 .use(loginRoutes)
 .use(statsRoutes)
+.use(exportRoutes)
 
 //on ajoute la gestion des erreurs 404
 app.use(({res}) => {
