+++ backend/src/controllers/exports/exportControllers.js
@@ -1,6 +1,46 @@
+const { Parser } = require("json2csv") ; 
 const ExcelJS = require('exceljs');
-const { Visite, Raison, Service } = require('../../models/index');
+const {Visite, Raison, Service} = require("../../models/index") ; 
 
+
+// expots des données en csv
+exports.exportVisitesCSV = async (req, res) => {
+    try {
+        const visites = await Visite.findAll({
+            include: [Service, Raison]
+        });
+  
+        const visitesData = visites.map(visite => ({
+            Id: visite.id,
+            Date: new Date(visite.date).toLocaleString('fr-FR', {
+                day: '2-digit',
+                month: '2-digit',
+                year: 'numeric',
+                hour: '2-digit',
+                minute: '2-digit'
+            }),
+            Email: visite.email,
+            Service: visite.service.nom,
+            Raison: visite.raison.typeRaison,
+            Satisfaction: visite.satisfaction,
+            Commentaire: visite.commentaire
+        }));
+  
+        const parser = new Parser();
+        const csv = parser.parse(visitesData);
+    
+        res.header('Content-Type', 'text/csv');
+        res.attachment('visites.csv');
+        res.send(csv);
+    } catch (error) {
+
+        console.error('Erreur lors de l\'exportation CSV :', error); 
+
+        res.status(500).json({ message: 'Erreur lors de l\'exportation CSV', error });
+    }
+};
+
+// export des données en excel
 exports.exportVisitesExcel = async (req, res) => {
     try {
         const visites = await Visite.findAll({
@@ -58,3 +98,4 @@ exports.exportVisitesExcel = async (req, res) => {
         res.status(500).json({ message: "Erreur lors de l'exportation Excel", error });
     }
 };
+
